Add hidden form elements
The form is submitted using JavaScript in static/search.js, but it
depends on being able to bind actions to form elements which we've
commented out. This brings the elements back as hiddent input elements.
I've also removed the commented out HTML. On new releases of pywb we
will likely have some work to do to port our changes forward. But that
will probably be easiest to do by modifying the default form that is
provided by pywb.

Fixes #118