Google Analytics 4
Add the new Google Analytics 4 embed code. pywb has two header
templates, one that is used for regular pages (head.html) and one that
is used when viewing archived content (banner.html).

An experimental event is also sent with the hostname of the archived
website to see if it is helpful in tracking what web content is being
viewed in SWAP.

Refs #158