Migrate to pywb 2.8
These are changes to upgrade to the latest version of pywb 2.8. It
doesn't look like there are any changes to the templates we've
overridden, mostly just changes to the VueJS application. One change I
noticed is the addition of a Print icon to the right in the menu bar,
which allows you to print an archived web page without the boilerplate.

pywb 2.8 requires Python v3.11, so this PR includes a small change to
the deploy to build the poetry virtual environment with Python v3.11
that needs to be installed (via puppet).

See https://github.com/sul-dlss/operations-tasks/issues/3742