Enable website stats gathering
It has been helpful to collect statistics about what web content is
being viewed in SWAP. To do that we need to determine what URL is being
viewed in the archive, extract the hostname from that, and then send it
along to Google Analytics as a custom event.

Previously we were using the Jinja2 `url` variable supplied by pywb. But
we can't (currently) interpolate that into JavaScript without
introducing a potential XSS vulnerability. This solution is to simply
look for the URL that's embedded in the `window.location.pathname` and
use that instead.